feat: add onboarding faucet step
