+++ src/components/@shared/Onboarding/Steps/Faucet.tsx
@@ -1,0 +1,106 @@
+import { LoggerInstance } from '@oceanprotocol/lib'
+import { getErrorMessage } from '@utils/onboarding'
+import { ReactElement, useEffect, useState } from 'react'
+import { toast } from 'react-toastify'
+import { useAccount, useNetwork, useProvider, useSignMessage } from 'wagmi'
+import { OnboardingStep } from '..'
+import { getSupportedChainIds } from '../../../../../chains.config'
+import content from '../../../../../content/onboarding/steps/importCustomTokens.json'
+import { getNonce, requestTokens } from '../../../../@utils/faucet'
+import StepBody from '../StepBody'
+import StepHeader from '../StepHeader'
+
+export default function Faucet(): ReactElement {
+  const { title, subtitle, body, image }: OnboardingStep = content
+
+  const { address: accountId } = useAccount()
+  const web3Provider = useProvider()
+  const { chain } = useNetwork()
+
+  const {
+    data: signMessageData,
+    error: signMessageError,
+    isLoading: signMessageLoading,
+    signMessage
+  } = useSignMessage()
+
+  const [loading, setLoading] = useState(false)
+  const [completed, setCompleted] = useState(false)
+
+  const faucetTokenRequest = async () => {
+    try {
+      await requestTokens(accountId, signMessageData)
+      setCompleted(true)
+    } catch (error) {
+      toast.error('Unable to request tokens. Please try again.')
+      LoggerInstance.error('[Onboarding] Error requesting tokens', error)
+    } finally {
+      setLoading(false)
+    }
+  }
+
+  const getAndSignNonce = async () => {
+    setLoading(true)
+    try {
+      if (!getSupportedChainIds().includes(chain?.id))
+        throw new Error(
+          'The chain you are connected to with your wallet is not supported'
+        )
+      LoggerInstance.log('[Onboarding] Requesting nonce from faucet', {
+        accountId
+      })
+      const nonce = await getNonce(accountId)
+
+      signMessage({ message: nonce.toString() })
+    } catch (error) {
+      toast.error(
+        getErrorMessage({
+          accountId,
+          web3Provider: !!web3Provider,
+          networkId: chain?.id
+        })
+      )
+      if (error.message) console.error(error.message)
+
+      setLoading(false)
+    }
+  }
+
+  useEffect(() => {
+    if (signMessageLoading) return
+
+    if (signMessageError) {
+      toast.error('Unable to sign message. Please try again.')
+      LoggerInstance.error(
+        '[Onboarding] Error signing message',
+        signMessageError
+      )
+      return
+    }
+
+    if (signMessageData) {
+      faucetTokenRequest()
+    }
+
+    setCompleted(true)
+  }, [signMessageData, signMessageError, signMessageLoading])
+
+  const actions = [
+    {
+      buttonLabel: `Request Test EUROe Tokens`,
+      buttonAction: async () => {
+        await getAndSignNonce()
+      },
+      successMessage: `Successfully requested test tokens.`,
+      loading,
+      completed
+    }
+  ]
+
+  return (
+    <div>
+      <StepHeader title={title} subtitle={subtitle} />
+      <StepBody body={body} image={image} actions={actions} />
+    </div>
+  )
+}
